requirements.txt
# Core Dependencies
tensorflow>=2.10.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.1.0
pillow>=9.0.0

# Web Interface & API
gradio>=3.50.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Logging & Monitoring
tensorboard>=2.10.0
tqdm>=4.65.0

# Data Processing
opencv-python>=4.7.0
pandas>=1.5.0

# Model Export
tf2onnx>=1.15.0
onnx>=1.14.0

# Optional (for advanced features)
# jupyter>=1.0.0
# notebook>=7.0.0
